+++ src/network/terrier_server.cpp
@@ -88,6 +88,6 @@ void TerrierServer::Close() {
 /**
  * Change port to new_port
  */
-void TerrierServer::SetPort(int new_port) { port_ = new_port; }
+void TerrierServer::SetPort(uint16_t new_port) { port_ = new_port; }
 
 }  // namespace terrier::network
